Rename IDMEFMessage into AbstractIDMEFMessage
Also, add a new IDMEFMessage class that acts as a list container
for a collection of IDMEF messages.
This container is not visible from the message objects themselves
as it is not formally listed as the messages' parent.

Meanwhile, the new class makes it easier to unserialize XML-encoded
IDMEF messages because the class hierarchy now follows the XML DOM
more closely.