Fix an issue with list offsets
An issue was fixed whereby it wasn't previously possible to set a value
using a direct offset ($alert->source[1]->...), even though the offset
matched the list's current length.

The expected behaviour in that case was to append the new value at the
end of the list. Instead, the first value of the list was silently
overwritten.